POST JSON request with user email.
Added some changes in the attemptPost method on the Main Activity class. The post method now properly sends the header, and the json object with the email content is converted in a entity(Byte array entity). In previous versions the API was returning the 400 Bad Request status code, now its returning 200.