UserDataModel constructor changes.
Added some changes to UserDataModel constructor. Added the correct method to get the user JSONObject and get the token string. Added a getmToken() method to the constructor so the model can return the data. 

In the MainActivity added the storeToken() method to store the token locally. Added the intent to navigate to the next string when the user token is successfully stored.