Fetch images from API and display in the list.
Added a new feature in ListActivity.Java to create a list o dogDataModel, set its url and pass it through the adapter.  In the Recyclerview adapter, added the Glide load function to set the image in the Image View.  Added a new blank constructor on DogDataModel and a setImgUrl method.